rename math to _math per #8041
