+++ bindings/pydrake/BUILD.bazel
@@ -83,7 +83,7 @@ drake_py_library(
 
 drake_pybind_library(
     name = "math_py",
-    cc_so_name = "math",
+    cc_so_name = "_math",
     cc_srcs = ["math_py.cc"],
     package_info = PACKAGE_INFO,
     py_deps = [
