+++ bindings/pydrake/systems/test/controllers_test.py
@@ -11,7 +11,7 @@ import unittest
 
 from pydrake.examples.pendulum import PendulumPlant
 from pydrake.systems.analysis import Simulator
-from pydrake.math import BarycentricMesh
+from pydrake._math import BarycentricMesh
 from pydrake.systems.controllers import (
     DynamicProgrammingOptions, FittedValueIteration)
